Revert "60fps volmeters (#2559)"
This reverts commit b58c1fab4774171dcdf317716579f828fe7be3c5.
