+++ app/components/MixerItem.vue
@@ -9,7 +9,7 @@
     </div>
   </div>
 
-  <MixerVolmeter :audioSource="props.audioSource" :volmetersEnabled="props.volmetersEnabled" v-if="!performanceMode"></MixerVolmeter>
+  <MixerVolmeter :audioSource="props.audioSource" v-if="!performanceMode"></MixerVolmeter>
 
   <div class="flex">
     <slider-input
