+++ app/components/MixerItem.vue.ts
@@ -11,7 +11,6 @@ import { SourcesService } from 'services/sources';
 
 class MixerItemProps {
   audioSource: AudioSource = null;
-  volmetersEnabled = false;
 }
 
 @Component({
