+++ app/components/pages/onboarding-steps/HardwareSetup.tsx
@@ -92,7 +92,6 @@ export default class HardwareSetup extends TsxComponent {
             >
               <MixerVolmeter
                 audioSource={this.defaultHardwareService.selectedAudioSource}
-                volmetersEnabled={true}
                 class={styles.volmeterCenter}
               />
             </div>
